Change numberOfGrids back to int and cast to double in division
The Bug was just an accidental integer division.  But for the code to
be expressive, the type of a variable should represent what it stands
for, so I think that the int type is more natural.

+++ tracking/modules/trackTimeExtraction/src/FullGridTrackTimeExtractionModule.cc
@@ -242,7 +242,7 @@ void FullGridTrackTimeExtractionModule::event()
   tries.reserve(m_param_numberOfGrids * 4);
   std::vector<T0Try> convergedTries;
 
-  const double deltaT0 = 1 / m_param_numberOfGrids * (m_param_maximalT0Shift - m_param_minimalT0Shift);
+  const double deltaT0 = 1.0 / (double)m_param_numberOfGrids * (m_param_maximalT0Shift - m_param_minimalT0Shift);
 
   for (unsigned int i = 1; i < m_param_numberOfGrids; i++) {
     const std::vector<T0Try>& tmpTries = extractTrackTimeFrom(recoTracks, m_param_minimalT0Shift + i * deltaT0, 2,
